+++ WriteSTL.cpp
@@ -225,7 +225,7 @@ MBErrorCode WriteSTL::get_triangle_data( const double coords[9],
   n[0] = e1[1]*e2[2] - e1[2]*e2[1];
   n[1] = e1[2]*e2[0] - e1[0]*e2[2];
   n[2] = e1[0]*e2[1] - e1[1]*e2[0];
-  float inv_len = 1.0f / sqrtf( n[0]*n[0] + n[1]*n[1] + n[2]*n[2] );
+  float inv_len = 1.0f / sqrt( n[0]*n[0] + n[1]*n[1] + n[2]*n[2] );
   n[0] *= inv_len;
   n[1] *= inv_len;
   n[2] *= inv_len;
